Refactoring und beginn mit der Auth Sequenze implementierung.

+++ bestia-zoneserver/src/main/java/net/bestia/zoneserver/game/service/AccountService.java
@@ -35,7 +35,7 @@ public class AccountService extends Service {
 		this.account = account;
 	}
 
-	public int getAccountId() {
+	public long getAccountId() {
 		return account.getId();
 	}
